+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 - Require higher TYPO3 Core bugfix versions (#2852)
-- Switch to our own copy of `TemplateHelper` (#2833, #2836, #2838, #2840)
+- Switch to our own copy of `TemplateHelper` (#2833, #2836, #2838, #2840, #2855)
 - Allow installations with static_info_tables V12 (#2824)
 - Reduce usage of `AbstractPlugin` (#2811)
 - Do not require jQuery anymore (#2807)